refactor(reports): add PageHeader section separators and extract card components
