+++ client/src/components/pages/Store/Reports/Charts/index.js
@@ -1,3 +1,4 @@
+export { AnalyticsCard } from "./AnalyticsCard";
 export { PaymentMethodPieChart } from "./PaymentMethodPieChart";
 export { RevenueAreaChart } from "./RevenueAreaChart";
 export { TopProductsBarChart } from "./TopProductsBarChart";
