+++ client/src/components/pages/Store/Reports/Filters/index.js
@@ -1,1 +1,2 @@
 export { DateRangeCard } from "./DateRangeCard";
+export { FiltersCard } from "./FiltersCard";
