+++ client/src/components/pages/Store/Reports/Sales/index.js
@@ -1,2 +1,3 @@
 export { SalesCard } from "./SalesCard";
+export { SalesDetailCard } from "./SalesDetailCard";
 export { SalesList } from "./SalesList";
