+++ client/src/components/pages/Store/Reports/Summary/index.js
@@ -1,2 +1,3 @@
 export { ReportSkeleton } from "./ReportSkeleton";
+export { SummaryCard } from "./SummaryCard";
 export { SummaryStat } from "./SummaryStat";
